requirements.txt
asgiref==3.6.0
Django==4.1
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
PyYAML==6.0
six==1.16.0
slippers==0.5.0
sqlparse==0.4.4
tzdata==2023.3
